rename utilities to known names
