+++ src/utilities/index.ts
@@ -1,8 +1,8 @@
 export { isValid as isValidPIS } from './pis';
+export { isValid as isValidCep } from './cep';
 export { isValid as isValidPhone } from './phone';
 export { isValid as isValidEmail } from './email';
-export { isValid as isValidPostalCode } from './postal-code';
 export { format as formatJudicialProcess } from './judicial-process';
-export { format as formatBankSlip, isValid as isValidBankSlip } from './bank-slip';
+export { format as formatBoleto, isValid as isValidBoleto } from './boleto';
 export { format as formatCPF, generate as generateCPF, isValid as isValidCPF } from './cpf';
 export { format as formatCNPJ, generate as generateCNPJ, isValid as isValidCNPJ } from './cnpj';
